# backend/storage/api-docs-admin/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Next Starter API — Admin",
        "description": "Documentação da área **Admin** (equipe interna / back-office) do projeto Laravel Next Starter.\n\nEsta documentação é restrita: o acesso à interface Swagger UI e ao JSON da spec exige autenticação HTTP Basic própria da documentação (usuário/senha definidos em `.env`), **independente** do login da API.\n\nApós acessar a documentação, os endpoints em si continuam exigindo o Bearer JWT obtido em `POST /api/admin/login`.",
        "contact": {
            "name": "Equipe de Desenvolvimento"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API (relativo ao host onde o Swagger UI está sendo servido)"
        }
    ],
    "paths": {
...
        "/admin/2fa/habilitar": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Iniciar habilitação do 2FA",
                "description": "Gera o secret TOTP e a URL otpauth para configurar o app autenticador. O 2FA só é efetivamente habilitado após a confirmação do código em /2fa/confirmar.",
                "operationId": "a79faeb02611fff21a64aa85dd8e8e94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "senha"
                                ],
                                "properties": {
                                    "senha": {
                                        "description": "Senha atual do usuário, para confirmar a identidade antes de gerar o secret.",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dados para configurar o app autenticador.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AutenticacaoDoisFatoresHabilitarResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BusinessError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/2fa/confirmar": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Confirmar habilitação do 2FA",
                "description": "Confirma o código gerado pelo app autenticador, efetivando a habilitação do 2FA para o usuário.",
                "operationId": "67605e97540c09b997b209bd966133f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codigo"
                                ],
                                "properties": {
                                    "codigo": {
                                        "type": "string",
                                        "pattern": "^[0-9]{6}$",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA ativado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AcaoMensagemResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BusinessError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/2fa/desabilitar": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Desabilitar 2FA",
                "description": "Desabilita o 2FA do usuário autenticado, exigindo a senha atual e um código válido do app autenticador.",
                "operationId": "e9125738f8f0b9c5f8d4755413d89ce7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "senha",
                                    "codigo"
                                ],
                                "properties": {
                                    "senha": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "pattern": "^[0-9]{6}$",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "2FA desabilitado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AcaoMensagemResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BusinessError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationErrorResponse": {
                "properties": {
                    "errors": {
                        "description": "Formato de erro de validação (422), produzido pelo handler de exceções\nglobal (bootstrap/app.php) a partir de uma ValidationException.",
                        "type": "object",
                        "example": {
                            "campo": [
                                "A mensagem de validação correspondente."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "BusinessErrorResponse": {
                "properties": {
                    "errors": {
                        "properties": {
                            "business": {
                                "description": "Formato padrão de erro de negócio/HTTP (400/401/403/404/500), produzido\npelo mesmo handler global para BusinessException, AuthorizationException,\nModelNotFoundException, HttpExceptionInterface, etc.",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Descrição do erro de negócio."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "AcaoMensagemResponse": {
                "description": "Resposta simples de confirmação para ações pontuais (ex.: habilitar/confirmar/desabilitar 2FA).",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Ação realizada com sucesso."
                    }
                },
                "type": "object"
            },
            "AutenticacaoDoisFatoresHabilitarResponse": {
                "description": "Dados para configurar o app autenticador (ex.: Google Authenticator): o secret pode ser digitado manualmente e o otpauth_url pode ser transformado em QR Code no front-end.",
                "properties": {
                    "secret": {
                        "type": "string",
                        "example": "JBSWY3DPEHPK3PXP"
                    },
                    "otpauth_url": {
                        "type": "string",
                        "example": "otpauth://totp/LaravelNextStarter:usuario@empresa.com?secret=JBSWY3DPEHPK3PXP&issuer=LaravelNextStarter"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "ValidationError": {
                "description": "Erro de validação dos campos enviados.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorResponse"
                        }
                    }
                }
            },
            "BusinessError": {
                "description": "Erro de regra de negócio (ex.: violação de uma regra da aplicação).",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Não autenticado: token ausente, inválido, expirado, ou sessão associada ao token não está mais ativa.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via **JSON Web Token (JWT)**.\n\n1. Faça login em `POST /api/login` (Private) ou `POST /api/admin/login` (Admin).\n2. Envie o token retornado no header `Authorization: Bearer {token}` nas demais requisições.\nO token carrega um `session_id`, validado a cada requisição contra a sessão ativa do usuário (tabela de sessões).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}